refactor(sft): align module name with track builder role
