+++ examples/sft/qwen_vl_sft.yaml
@@ -91,7 +91,7 @@ stack:
   max_grad_norm: 1.0
 
 track_builder:
-  _target_: unirl.train.sft.source.ARSupervisedTrackBuilder
+  _target_: unirl.train.sft.track_builder.ARSupervisedTrackBuilder
   max_response_length: 1024
 
 data_source:
