+++ examples/sft/validation/qwen_vl_sft_lora.yaml
@@ -90,7 +90,7 @@ stack:
   max_grad_norm: 1.0
 
 track_builder:
-  _target_: unirl.train.sft.source.ARSupervisedTrackBuilder
+  _target_: unirl.train.sft.track_builder.ARSupervisedTrackBuilder
   max_response_length: 1024
 
 data_source:
