+++ examples/sft/validation/sd3_sft_lora.yaml
@@ -98,7 +98,7 @@ stack:
   max_grad_norm: 1.0
 
 track_builder:
-  _target_: unirl.train.sft.source.DiffusionSupervisedTrackBuilder
+  _target_: unirl.train.sft.track_builder.DiffusionSupervisedTrackBuilder
   height: 512
   width: 512
   guidance_scale: 1.0
