+++ unirl/train/sft/__init__.py
@@ -5,7 +5,7 @@ driver is ``unirl/trainer/sft.py``; this package holds only the piece that is
 genuinely new to supervision: turning dataset records into stage-ready tracks.
 """
 
-from unirl.train.sft.source import (
+from unirl.train.sft.track_builder import (
     ARSupervisedTrackBuilder,
     DiffusionSupervisedTrackBuilder,
     SupervisedTrackBuilder,
